Extract StrBuilderTokenizer to its own file
Improves maintainability by following the Single Responsibility Principle.
StrBuilder is reduced from 3089 to ~3060 lines.
No functional changes — all existing behavior preserved.
